+++ modules/AWSPowerShell/AWSPowerShellCompleters.psm1
@@ -3013,7 +3013,7 @@ $AAR_Completers = {
             ($_ -eq "Update-AARService/SourceConfiguration_CodeRepository_CodeConfiguration_CodeConfigurationValues_Runtime")
         }
         {
-            $v = "NODEJS_12","PYTHON_3"
+            $v = "CORRETTO_11","CORRETTO_8","NODEJS_12","NODEJS_14","PYTHON_3"
             break
         }
 
@@ -47295,6 +47295,16 @@ $TRN_Completers = {
             break
         }
 
+        # Amazon.Translate.Formality
+        {
+            ($_ -eq "ConvertTo-TRNTargetLanguage/Settings_Formality") -Or
+            ($_ -eq "Start-TRNTextTranslationJob/Settings_Formality")
+        }
+        {
+            $v = "FORMAL","INFORMAL"
+            break
+        }
+
         # Amazon.Translate.JobStatus
         "Get-TRNTextTranslationJobList/Filter_JobStatus"
         {
@@ -47353,6 +47363,7 @@ $TRN_map = @{
     "MergeStrategy"=@("Import-TRNTerminology")
     "OutputDataConfig_EncryptionKey_Type"=@("Start-TRNTextTranslationJob")
     "ParallelDataConfig_Format"=@("New-TRNParallelData","Update-TRNParallelData")
+    "Settings_Formality"=@("ConvertTo-TRNTargetLanguage","Start-TRNTextTranslationJob")
     "Settings_Profanity"=@("ConvertTo-TRNTargetLanguage","Start-TRNTextTranslationJob")
     "TerminologyData_Directionality"=@("Import-TRNTerminology")
     "TerminologyData_Format"=@("Import-TRNTerminology")