+++ modules/AWSPowerShell/Cmdlets/WAFV2/Basic/Write-WAF2LoggingConfiguration-Cmdlet.cs
@@ -29,11 +29,12 @@ namespace Amazon.PowerShell.Cmdlets.WAF2
 {
     /// <summary>
     /// Enables the specified <a>LoggingConfiguration</a>, to start logging from a web ACL,
-    /// according to the configuration provided.
+    /// according to the configuration provided. 
     /// 
-    ///  
-    /// <para>
-    /// You can access information about all traffic that WAF inspects using the following
+    ///  <note><para>
+    /// You can define one logging destination per web ACL.
+    /// </para></note><para>
+    /// You can access information about the traffic that WAF inspects using the following
     /// steps:
     /// </para><ol><li><para>
     /// Create your logging destination. You can use an Amazon CloudWatch Logs log group,
@@ -50,6 +51,9 @@ namespace Amazon.PowerShell.Cmdlets.WAF2
     /// the logging destination. For an Amazon CloudWatch Logs log group, WAF creates a resource
     /// policy on the log group. For an Amazon S3 bucket, WAF creates a bucket policy. For
     /// an Amazon Kinesis Data Firehose, WAF creates a service-linked role.
+    /// </para><para>
+    /// For additional information about web ACL logging, see <a href="https://docs.aws.amazon.com/waf/latest/developerguide/logging.html">Logging
+    /// web ACL traffic information</a> in the <i>WAF Developer Guide</i>.
     /// </para><note><para>
     /// This operation completely replaces the mutable specifications that you already have
     /// for the logging configuration with the ones that you provide to this call. To modify
@@ -95,8 +99,7 @@ namespace Amazon.PowerShell.Cmdlets.WAF2
         #region Parameter LoggingConfiguration_LogDestinationConfig
         /// <summary>
         /// <para>
-        /// <para>The Amazon Resource Names (ARNs) of the logging destinations that you want to associate
-        /// with the web ACL.</para>
+        /// <para>The logging destination configuration that you want to associate with the web ACL.</para><note><para>You can associate one logging destination to a web ACL.</para></note>
         /// </para>
         /// </summary>
         #if !MODULAR